+++ assets/javascripts/kitten/components/grid/lego-grid.js
@@ -1,10 +1,32 @@
 import React from 'react'
 import classNames from 'classnames'
 import Masonry from 'react-masonry-component'
+import { ScreenConfig } from 'kitten/constants/screen-config'
+import { NUM_COLUMNS } from 'kitten/constants/grid-config'
 
 export const LegoGrid = props => {
   const { className, masonryProps, children, ...others } = props
-  const gridClassName = classNames('k-LegoGrid', className)
+  const classByMediaQuery = () => {
+    const classNamesByMediaQuery = Object.keys(ScreenConfig).map(size => {
+      const mediaQuery = size.toLowerCase()
+      const items = props[`items-${mediaQuery}-up`]
+
+      return classNames(
+        classNamesByMediaQuery,
+        {
+          [`k-LegoGrid--${NUM_COLUMNS/items}col@${mediaQuery}`]: items,
+        },
+      )
+    })
+
+    return classNamesByMediaQuery
+  }
+
+  const gridClassName = classNames(
+    'k-LegoGrid',
+    classByMediaQuery(),
+    className,
+  )
 
   return (
     <div className={ gridClassName } { ...others }>
@@ -16,7 +38,10 @@ export const LegoGrid = props => {
 }
 
 LegoGrid.Item = ({ children, ...props }) => {
-  const itemClassName = classNames('k-LegoGrid__item', props.className)
+  const itemClassName = classNames(
+    'k-LegoGrid__item',
+    props.className,
+  )
 
   return (
     <div { ...props } className={ itemClassName }>