+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `items-sizeScreen-up` props to `LegoGrid` component.
 - Fix: Remove `textProps` props from `inputProps` Checkbox component.
 
 ## [20.3.0] - 2018-03-22